Refactor header structures in FPGA interchange Arch.
